add ability to save output to a file

+++ src/app.tsx
@@ -4,6 +4,7 @@ import { useState, useRef, useEffect, useMemo } from "react";
 import type { ProcessConfig } from "./args";
 import { spawnProcess, type CommandProcess, type LogEntry } from "./process";
 import { colorForCmd } from "./util";
+import { UncontrolledTextInput } from "ink-text-input";
 
 export interface AppProps {
   processConfigs: ProcessConfig[];
@@ -14,13 +15,13 @@ export const App: React.FC<AppProps> = ({ processConfigs }) => {
   const [scrollOffset, setScrollOffset] = useState(0);
   const [filter, setFilter] = useState<number | null>(null);
   const [selectedButton, setSelectedButton] = useState(0);
+  const [saveModal, setSaveModal] = useState(false);
   const { exit } = useApp();
   const processesRef = useRef<CommandProcess[]>([]);
   const [numColumns, numRows] = useStdoutDimensions();
+  const [error, setError] = useState("");
   const prevSelectedButton = useRef(selectedButton);
-
-  const handleLogEntry = (entry: LogEntry) =>
-    setLogs((prev) => [...prev, entry]);
+  const errorTimeout = useRef<NodeJS.Timeout | null>(null);
 
   // TODO -- wish this wasn't hardcoded.
   // The 4 is the sum of non-log-line UI elements in the output.
@@ -35,6 +36,23 @@ export const App: React.FC<AppProps> = ({ processConfigs }) => {
       ? logs
       : logs.filter((log) => log.command === filter);
   }, [logs, filter]);
+
+  const handleSaveFile = async (filename: string) => {
+    if (filename) {
+      try {
+        await Bun.write(filename, filteredLogs.map((l) => l.text).join("\n"));
+      } catch (e) {
+        setError((e as Error).message);
+      }
+    } else {
+      setError("Must specify a filename");
+    }
+
+    setSaveModal(false);
+  };
+
+  const handleLogEntry = (entry: LogEntry) =>
+    setLogs((prev) => [...prev, entry]);
 
   const prevFilteredLogs = useRef(filteredLogs);
 
@@ -55,6 +73,15 @@ export const App: React.FC<AppProps> = ({ processConfigs }) => {
   }
 
   useEffect(() => {
+    if (error) {
+      if (errorTimeout.current) {
+        clearTimeout(errorTimeout.current);
+      }
+      errorTimeout.current = setInterval(() => setError(""), 5000);
+    }
+  }, [error, setError]);
+
+  useEffect(() => {
     const processes: CommandProcess[] = processConfigs.map((proc) =>
       spawnProcess(proc, handleLogEntry)
     );
@@ -63,6 +90,18 @@ export const App: React.FC<AppProps> = ({ processConfigs }) => {
   }, [processConfigs]);
 
   useInput((input, key) => {
+    if (error) {
+      setError("");
+      return;
+    }
+
+    if (saveModal) {
+      if (key.escape) {
+        setSaveModal(false);
+      }
+      return;
+    }
+
     if (input === "q" || (key.ctrl && input === "c")) {
       processesRef.current.forEach(({ process }) => {
         process.kill();
@@ -77,6 +116,10 @@ export const App: React.FC<AppProps> = ({ processConfigs }) => {
           Object.assign(proc, spawnProcess(proc.config, handleLogEntry));
         }
       });
+    }
+
+    if (input === "s") {
+      setSaveModal(true);
     }
 
     if (key.upArrow) {
@@ -124,28 +167,43 @@ export const App: React.FC<AppProps> = ({ processConfigs }) => {
           ))
         )}
         <Spacer />
-        <Box gap={1}>
+        {error ? (
           <Box>
-            <Text inverse={selectedButton === 0} bold={filter === null}>
-              All
-            </Text>
+            <Text color="redBright">{error}</Text>
+            <Spacer />
+            <Text dimColor>[any key] hide error</Text>
           </Box>
-          {processConfigs.map((cmd) => (
-            <Box key={cmd.id}>
-              <Text
-                inverse={selectedButton === cmd.id + 1}
-                bold={filter === cmd.id}
-                color={colorForCmd(cmd.id)}
-              >
-                [{cmd.id}] {cmd.command.substring(0, 10)}
+        ) : saveModal ? (
+          <Box gap={1}>
+            <Text>Save to file:</Text>
+            <UncontrolledTextInput onSubmit={handleSaveFile} />
+            <Spacer />
+            <Text dimColor>[ret] submit, [esc] cancel</Text>
+          </Box>
+        ) : (
+          <Box gap={1}>
+            <Box>
+              <Text inverse={selectedButton === 0} bold={filter === null}>
+                All
               </Text>
             </Box>
-          ))}
-          <Spacer />
-          <Text dimColor>
-            [↑/↓] scroll, [←/→] filter, [q] quit, [r] restart proc(s)
-          </Text>
-        </Box>
+            {processConfigs.map((cmd) => (
+              <Box key={cmd.id}>
+                <Text
+                  inverse={selectedButton === cmd.id + 1}
+                  bold={filter === cmd.id}
+                  color={colorForCmd(cmd.id)}
+                >
+                  [{cmd.id}] {cmd.command.substring(0, 10)}
+                </Text>
+              </Box>
+            ))}
+            <Spacer />
+            <Text dimColor>
+              [↑/↓] scroll, [←/→] filter, [q] quit, [r] restart proc(s)
+            </Text>
+          </Box>
+        )}
       </Box>
     </Box>
   );
